Show shortlisting conditionally in menu with Delete
