+++ apps/app/src/components/decisions/ProposalCardActions.tsx
@@ -4,11 +4,7 @@ import { trpc } from '@op/api/client';
 import type { proposalEncoder } from '@op/api/encoders';
 import { ProfileRelationshipType } from '@op/api/encoders';
 import { Button } from '@op/ui/Button';
-import { DialogTrigger } from '@op/ui/Dialog';
-import { Modal, ModalBody, ModalFooter, ModalHeader } from '@op/ui/Modal';
-import { toast } from '@op/ui/Toast';
-import { Heart, Trash2 } from 'lucide-react';
-import { useState } from 'react';
+import { Heart } from 'lucide-react';
 import { LuBookmark } from 'react-icons/lu';
 import { z } from 'zod';
 
@@ -23,7 +19,6 @@ export function ProposalCardActions({
 }) {
   const t = useTranslations();
   const utils = trpc.useUtils();
-  const [isDeleteModalOpen, setIsDeleteModalOpen] = useState(false);
 
   // Subscribe to the individual proposal data which gets optimistically updated
   const { data: currentProposal } = trpc.decision.getProposal.useQuery(
@@ -270,31 +265,9 @@ export function ProposalCardActions({
       },
     });
 
-  const deleteProposalMutation = trpc.decision.deleteProposal.useMutation({
-    onError: (error, _variables) => {
-      toast.error({
-        message: error.message || t('Failed to delete proposal'),
-      });
-    },
-    onSuccess: () => {
-      toast.success({
-        message: t('Proposal deleted successfully'),
-      });
-    },
-    onSettled: () => {
-      // Always refetch after error or success
-      if (initialProposal.processInstance?.id) {
-        utils.decision.listProposals.invalidate({
-          processInstanceId: initialProposal.processInstance.id,
-        });
-      }
-    },
-  });
-
   const isLoading =
     addRelationshipMutation.isPending ||
-    removeRelationshipMutation.isPending ||
-    deleteProposalMutation.isPending;
+    removeRelationshipMutation.isPending;
 
   const handleLikeClick = async () => {
     if (!currentProposal.profileId) {
@@ -344,22 +317,6 @@ export function ProposalCardActions({
     }
   };
 
-  const handleDeleteConfirm = async () => {
-    if (!currentProposal.id) {
-      console.error('No proposal ID provided for delete action');
-      return;
-    }
-
-    try {
-      await deleteProposalMutation.mutateAsync({
-        proposalId: currentProposal.id,
-      });
-      setIsDeleteModalOpen(false); // Close modal after successful deletion
-    } catch (error) {
-      console.error('Error in ProposalCardActions handleDeleteConfirm:', error);
-    }
-  };
-
   return (
     <div className="flex w-full items-center gap-2 sm:w-auto">
       <Button
@@ -382,55 +339,6 @@ export function ProposalCardActions({
         <LuBookmark className="size-4" />
         {currentProposal.isFollowedByUser ? t('Following') : t('Follow')}
       </Button>
-      {false && initialProposal.isEditable && (
-        <DialogTrigger
-          isOpen={isDeleteModalOpen}
-          onOpenChange={setIsDeleteModalOpen}
-        >
-          <Button
-            size="small"
-            color="secondary"
-            className="w-full text-nowrap"
-            isDisabled={isLoading}
-          >
-            <Trash2 className="size-4" />
-            {t('Delete')}
-          </Button>
-          <Modal
-            isDismissable
-            isOpen={isDeleteModalOpen}
-            onOpenChange={setIsDeleteModalOpen}
-          >
-            <ModalHeader>{t('Delete Proposal')}</ModalHeader>
-            <ModalBody>
-              <p>
-                {t(
-                  'Are you sure you want to delete this proposal? This action cannot be undone.',
-                )}
-              </p>
-            </ModalBody>
-            <ModalFooter>
-              <Button
-                color="secondary"
-                className="w-full sm:w-fit"
-                onPress={() => setIsDeleteModalOpen(false)}
-              >
-                {t('Cancel')}
-              </Button>
-              <Button
-                color="destructive"
-                onPress={handleDeleteConfirm}
-                className="w-full sm:w-fit"
-                isDisabled={deleteProposalMutation.isPending}
-              >
-                {deleteProposalMutation.isPending
-                  ? t('Deleting...')
-                  : t('Delete')}
-              </Button>
-            </ModalFooter>
-          </Modal>
-        </DialogTrigger>
-      )}
     </div>
   );
 }
